app/requirements.txt
Flask
Flask-Cors
Flask_RESTFul
Flask_JWT_Extended
Flask_SQLAlchemy
Flask-PyMongo
Flask-CeleryExt
pymysql
python-dotenv
pandas
numpy
sklearn
celery
mysqlclient
requests
pymongo
statsmodels
apyori
